Simplify config reading logic
Fix #4

This simplifies the logic to read default and user configurations a bit
and resolves a bug if no user settings file exists yet. The 'settings'
dicts from default and user settings are merged, so that users can
overwrite a single entry without losing all the default values.
Additionally, the starting command for the server will be updated if
"sysimage_path" is present in the user settings, so that the server will
be started accordingly, in case that setting has been modified manually.