Remove files for precompilation
Julia 1.6 reduces the startup time for `using` packages, so the speedup
from precompiling the language server is not as huge anymore. Since
precompiling takes quite a while, creates a big sysimage file and is
required to be performed again for every update of the server and every
minor Julia version update, maybe it is not worth it to keep the
precompilation strategy. Also it would require to manage the adjusted
startup command with the sysimage path in the user settings.

In case the option to precompile the server is desired to be added back
later (with a better solution where to store the sysimage file), the
starting command for the server in precompile_execution_file.jl should
be adjusted to use the `runserver()` function.