Show documentation sheet after command if in background
When running the "Show Documentation" command from the context menu,
and there is already a documentation sheet but it is not visible, add it
to the selected sheets (side by side mode). Otherwise the user would get
no feedback that the command worked, and would need to find and focus
the documentation sheet manually.