Fix compatibility with Julia 1.12 and disable TestItem feature
This updates the language server to the latest version which is required
for compatibility with Julia 1.12.
Unfortunately the TestItem feature had to be disabled, because there
were a lot of internal API changes in LanguageServer.jl related to the
TestItems, which made it impossible to keep using the current
implementation.